rename icons folder to 'SocialstreamIcons'
