+++ stubs/inertia/resources/js/Components/SocialstreamIcons/BitbucketIcon.vue
