+++ stubs/inertia/resources/js/Components/SocialstreamIcons/FacebookIcon.vue
