+++ stubs/inertia/resources/js/Components/SocialstreamIcons/GitLabIcon.vue
