+++ stubs/inertia/resources/js/Components/SocialstreamIcons/GithubIcon.vue
