+++ stubs/inertia/resources/js/Components/SocialstreamIcons/GoogleIcon.vue
